+++ packages/basic/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jevon617/eslint-config-basic
 
+## 0.0.3
+
+### Patch Changes
+
+- close import/no-unresolved
+
 ## 0.0.2
 
 ### Patch Changes