+++ packages/typescript/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jevon617/eslint-config-ts
 
+## 0.0.3
+
+### Patch Changes
+
+- close import/no-unresolved
+- Updated dependencies
+  - @jevon617/eslint-config-basic@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes